Revert "[개선] User / Admin 서비스 로직 분리"
