--- src/main/java/com/bamdoliro/maru/shared/auth/annotation/RoleCheck.java
@@ -1,14 +1,0 @@
-package com.bamdoliro.maru.shared.auth.annotation;
-
-import com.bamdoliro.maru.shared.auth.Authority;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD,ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RoleCheck {
-    Authority[] value() default{};
-}
