+++ src/test/java/com/bamdoliro/maru/shared/util/ControllerTest.java
@@ -20,10 +20,8 @@ import com.bamdoliro.maru.infrastructure.neis.SearchSchoolService;
 import com.bamdoliro.maru.presentation.analysis.AnalysisController;
 import com.bamdoliro.maru.presentation.auth.AuthController;
 import com.bamdoliro.maru.presentation.fair.FairController;
-import com.bamdoliro.maru.presentation.form.AdminFormController;
 import com.bamdoliro.maru.presentation.form.DraftFormController;
 import com.bamdoliro.maru.presentation.form.FormController;
-import com.bamdoliro.maru.presentation.form.UserFormController;
 import com.bamdoliro.maru.presentation.message.MessageController;
 import com.bamdoliro.maru.presentation.notice.NoticeController;
 import com.bamdoliro.maru.presentation.question.QuestionController;
@@ -48,8 +46,6 @@ import org.springframework.test.web.servlet.MockMvc;
             SchoolController.class,
             QuestionController.class,
             FormController.class,
-            AdminFormController.class,
-            UserFormController.class,
             NoticeController.class,
             DraftFormController.class,
             FairController.class,
